Moved rfm_funcs up a level
