+++ src/server/secrets/kustomization.yaml
@@ -1,0 +1,17 @@
+secretGenerator:
+
+- name: db-pass
+  literals:
+  - POSTGRES_PASSWORD="thispasswordisverysecure"
+
+- name: api-server-pws
+  literals:
+  - BASEUSER_PW="d74jdHRZi&H@k;<?{_=+384$5"
+  - BASEEDITOR_PW="LUH%547865HKG7765]][-7653] KYFG kjg"
+  - BASEADMIN_PW="asdf747^*%&HLKUH Re4UF:IOUUy((*(#@#$9R9"
+  - JWT_SECRET="wK[d6W__njZx>T_4uZ2Hb[NwJvf}>s4j1rVaB0TeCbmltmvx8]70zYP1HhVZ5]GK"
+  - APP_SECRET_KEY="]t_HNOfWWNiT1qYaX>zGmesQ<WgpBV4b1Dv51Ks98P>Lppp9A6IIeiC{clXSm6hg"
+  - SHELTERLUV_SECRET_TOKEN="d71f9b6e-e70d-45a8-a695-5e86932ddd6e"
+
+generatorOptions:
+  disableNameSuffixHash: true
